Better config defaulting, provide access to pool and server
Two changes in here that probably should have been separate commits. The
first is to have the `start-server` function handle configuration
defaults better. Previously, this code:

```clojure
(start-server {:queues [{:name "default"}]})
```

would fail because the `:num-workers` property in the configuration was
not set. This is due to `start-server` expecting a fully-fleshed out
configuration. Now, it merges its first argument with a config loaded in
the `config` ns.

The second change is to store the most recent server and Jedis pool in
global atoms. This is a convenience that allows one to do stuff like:

```clojure
(start-server)
(enqueue {:fn-var #'my-fn} @pool*)
```

I am not sure whether an atom is really necessary here. I may just
switch to putting `(def pool* pool)` in `start-server`.