New arities for enqueue and stop-server to use defaults
This changes the enqueue and stop-server functions to default to using
the values in the pool* and server* atoms. For most people, this will be
sufficient for all their needs, so it makes things easier to use.

Theoretically this could be done for all the functions that accept a
pool, but I think that's overkill and will be a lot of unnecessary code.
Those functions are not part of the public API.