Rewrite threading logic to use a thread pool
This uses a fixed thread pool instead of creating clojure futures. There
are a few benefits to doing this:

- The existing implementation requires the use of `shutdown-agents` to
get applications to shut down.
- The code is a little simpler now.
- There is control over how long we will wait for the pool to shut down.
If a user wants the pool to only wait a few seconds and then shut down
more abruptly, this can be done.

Notably there is no handling currently when the thread does not shut
down all the way. This should be handled in some way.