Redo config management to eliminate big config map
The main reason this changed happened is for the Farmhand UI code. The
`load-config` function returned a map containing things like
`:num-workers`, which is just not relevant to the UI. The `:redis` key,
however, IS relevant to the UI. This has always made a bit
uncomfortable.

However, there is a practical issue at stake as well. The UI needs to
provide its own custom configuration values and there is no way for it
to easily do so, since `load-config` takes no options.

After trying a few different ways of implementing the configuration
code, the result in here is my favorite. It is very flexible and
extendable.