Redo handler namespace to simplify default middleware
The number of middleware functions was a little out of control. It means
if you wanted to put your own middleware in the midst of it all, it
would require a lot of effort.

In fact I'm not even sure this change goes far enough. The `wrap-outer`
function should maybe just be part of the `work` namespace. Or perhaps
added by default within the `start-server` function. OTOH doing this
does make things less flexible from the user's point of view. So simple
or flexible?