New middleware for automatic job retry
This implements a feature I've been wanting to get to for awhile:
automatic job retries. There is new middleware that makes this happen.

To invoke it (at least as of this commit), you would just have to
enqueue your jobs with

```clojure
(enqueue {:fn-var #'job-fn :retry {:strategy "backoff"}})
```

It is extensible via a multimethod that dispatches based on the retry
strategy. So someone could write their own retry strategy with minimal
work.