Refactor of Jedis macros
This is inspired by how clojure JDBC works. This issue:
https://github.com/b-ryan/farmhand/issues/3 is what eventually led to
this idea.

The basic idea here is that the `with-jedis` and `with-transaction`
macros should be a little more generic. They take and pass around a
"context" map which may have a transaction or Jedis object inside it.
Then when you need to do low-level Jedis operations you extract the
specific thing you need.

This setup allows nesting things nicely. It will remove the need for
duplicate versions of functions where one takes a pool and one takes a
specific class.