Smarter data cleanup
This implements the bulk of the smarter data cleanup discussed in
https://github.com/b-ryan/farmhand/issues/2

TTL can be customized (untested)

There is now a single point of truth for when a job will expire. There
is no TTL put on the job record itself, but instead the job is cleaned
up when the job is removed from a registry.

As a result of the above, the TTL will refresh whenever a job is moved
from one registry to another (ie. when it is requeued the job data will
not expire).

Still to do, however, is to consolidate the way jobs are removed from
registries. For example, when a job is removed from the in progress
registry, it should be put into the failed registry. And the scheduling
code could reuse the same logic in this ns.