Gigantic rework of schedule/registry code
This consolidates a lot of code related to registries and scheduling and
makes it possible to taken different actions depending on the registry
being acted upon.

The scheduling code as a result is able to take full advantage of this
to eliminate its own version of the same code.

Also during the process, we now semi-properly handle in-flight jobs that
expire. They will be moved to the dead letter queue when they expire
rather than being deleted.