Big reworking of job handling throughout the processing code
The original goal was to have the "complete" and "fail" functions in the
queue namespace be able to respect a user-defined TTL on the result of
the job. In order to do this they need to know more than just the job-id
of job.

I also have been dissatisfied with the way jobs are treated in the
handler ns. Properties of the job are updated in Redis, but the changes
are not reflected in the in-memory version of the job.

This changes the code so it updates the job whenever properties are
modified.