Gigantic, not-quite-working refactor
Ultimately this change centers around one thing: job management

Due to the middleware design, it is really error-prone and tricky to
simultaneously update Redis with job information (like the status,
retry, etc) and keep it in sync with the in-memory job.

This change centralizes things by having the code that actually talks to
redis centralized in the wrap-outer function. This means there will be a
consistent view of the job during the whole processing lifecycle.