Redo serialization of data to Redis
I don't want to rely too much on EDN as the data storage. At the least,
I want an easy way to start using other serializers, for example if I
wanted to extend this project to Java I would want to use some Java
object serialization for the jobs & arguments. Mostly I want some
high-level properties on the job object to just be stored as strings
with simpler parsing on the dequeue-side.