Redo retry storage so that state is separate from definition
This changes up the way retries are done. Instead of modifying the
:retry map, retries can modify the [:retry :state] path of the job. This
allows for a distinction between the "state" of the retry and the
definition of the retry. When the retry is done, [:retry :state] can be
set to nil, but the definition itself will remain intact. This means
when requeuing a job, it retries itself the same way.